pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.5.2",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "uvicorn[standard]>=0.41.0",
]
